requirements.txt
appdirs==1.4.3
Django==1.11
django-filter==1.0.2
django-jinja==2.3.0
django-mysql==1.1.1
djangorestframework==3.6.2
ez-setup==0.9
gunicorn==19.7.1
Jinja2==2.9.6
MarkupSafe==1.0
mysqlclient==1.3.10
packaging==16.8
psycopg2==2.7.1
pyparsing==2.2.0
pytz==2017.2
six==1.10.0
